00045: connected to new backend function to prevent two user logins for same account
