+++ src/app/(home)/brothers/library/[username]/page.tsx
@@ -148,6 +148,7 @@ const Dashboard: React.FC = () => {
     getData();
     getBookData();
   }, []);
+  console.log(findUser);
   return (
     <>
       {loading ? (
