+++ src/components/header/HeaderCon.tsx
@@ -4,8 +4,7 @@
 import Login from "../user/Login";
 import HeaderLinks from "./HeaderLinks";
 import HeaderLinksMobile from "./HeaderLinksMobile";
-import { Button } from "antd";
-import { useContext, useState, useRef } from "react";
+import { useContext, useRef } from "react";
 import { AuthContext } from "../../context/authContext";
 import { useJwt } from "react-jwt";
 import { getTime } from "../../functions/timeFunction";
@@ -13,25 +12,16 @@ import Link from "next/link";
 import { useMediaQuery } from "react-responsive";
 import style from "./HeaderCob.module.css";
 import Logo from "../misc/Logo";
+import Logout from "../user/Logout";
 
 const HeaderCon: React.FC = () => {
-  const [loadings, setLoadings] = useState([]);
-  const { logout, token } = useContext(AuthContext);
+  const { token } = useContext(AuthContext);
   const { decodedToken }: { decodedToken?: { token: string, username: string } } =
     useJwt(token);
 
   const handleDesktop = useMediaQuery({ query: "(min-device-width: 801px)" });
   const headerCon = useRef(null);
 
-  const handleClick = () => {
-    setLoadings([true]);
-    setTimeout(() => {
-      localStorage.removeItem("token");
-      localStorage.removeItem("username");
-      setLoadings([false]);
-      logout();
-    }, 5000);
-  };
   const headerMessage = getTime();
 
   //neccessary for the (3d) layout
@@ -40,18 +30,7 @@ const HeaderCon: React.FC = () => {
     <header ref={headerCon} className={token ? style.headerConToken : style.headerConNoToken}>
       {token ? (
         <>
-          <div className="flex items-center">
-            <Button
-              className="m-5"
-              type="primary"
-              ghost
-              loading={loadings[0]}
-              onClick={handleClick}
-            >
-              Logout
-            </Button>
-          </div>
-
+          <Logout />
           {handleDesktop ? (
             <div className={style.headerLinks}>
               <HeaderLinks />
