+++ src/components/user/LoginForm.module.css
@@ -1,0 +1,8 @@
+.errorH {
+    background-color: black;
+    color: red;
+    padding: 1rem;
+    font-weight: 900;
+    text-align: center;
+    margin-bottom: 1rem;
+}
